requirements.txt
beautifulsoup4
selenium
Pillow
pytest